+++ LanCloud/LanCloud/Domain/FileStripe/IFileStripe.cs
@@ -7,7 +7,7 @@ namespace LanCloud.Domain.FileStripe
         string Extention { get; }
         string Hash { get; }
         long Length { get; }
-        byte[] Indexes { get; }
+        int[] Indexes { get; }
         bool IsTemp { get; }
 
         FileStream OpenRead();
