+++ LanCloud/LanCloud/Domain/Share/IShareStripe.cs
@@ -3,6 +3,6 @@
     public interface IShareStripe
     {
         IShare Share { get; }
-        byte[] Indexes { get; }
+        int[] Indexes { get; }
     }
 }
